ENH: Change from Debug to Warning macro.
When non DICOM and non standard ITK tags are in the metadata
dictionary, issue a warning. This addresses the common mistake
of using the comma as a separator in the group-element key
instead of the pipe, which is what ITK/GDCM uses. These metadata
dictionary entries were silently ignored. The warning should provide
the user with a hint about what is happening.